+++ app/src/main/java/com/issueissyu/fe/data/remote/dto/response/collection/SetProfileResponse.kt
@@ -1,6 +1,6 @@
 package com.issueissyu.fe.data.remote.dto.response.collection
 
 data class SetProfileResponse(
-    val profileCollectionId: Int,
+    val profileCollectionId: Long,
     val profileImageUrl: String
 )
